Done, chưa responsive

+++ controllers/categoryCtrl.js
@@ -1,4 +1,6 @@
 const Category = require('../models/categoryModel')
+const Products = require('../models/productModel')
+
 
 const categoryCtrl = {
     getCategories: async(req, res) => {
@@ -27,6 +29,9 @@ const categoryCtrl = {
     },
     deleteCategory: async (req, res) => {
         try {
+            const products = await Products.findOne({category: req.params.id})
+            if(products) return res.status(400).json({msg: "Hãy xóa hết sản phẩm trong category này rồi thử lại"})
+
             await Category.findByIdAndDelete(req.params.id)
             res.json({msg: "Xóa category thành công"})
         } catch (err) {
